# server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Status": {
				"enum": [
					"ordering",
					"ordered",
					"delivery",
					"canceled",
					"completed"
				],
				"type": "string"
			},
			"PaymentMethod": {
				"enum": [
					"cod",
					"banktransfer",
					"e-wallet",
					"gateway"
				],
				"type": "string"
			},
...
			"Change": {
				"enum": [
					"DESC",
					"ASC"
				],
				"type": "string"
			},
			"UserRole": {
				"enum": [
					"all",
					"admin",
					"editor",
					"customer"
				],
				"type": "string"
			},
...
			"IOrderBasicProps": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"paymentCheck": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"shipping": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					},
					"orderAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status",
					"paymentMethod",
					"paymentCheck",
					"address",
					"total",
					"orderAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUserSearchProps": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"fname": {
						"type": "string"
					},
					"lname": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"sale": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"fname",
					"lname",
					"phone",
					"address",
					"email",
					"role",
					"createdAt",
					"sale"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUsers": {
				"properties": {
					"users": {
						"items": {
							"$ref": "#/components/schemas/IUserSearchProps"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"users",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserField": {
				"enum": [
					"fname",
					"sale",
					"createdAt"
				],
				"type": "string"
			},
			"IUserWithOrders": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"fname": {
						"type": "string"
					},
					"lname": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/IOrderBasicProps"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"fname",
					"lname",
					"phone",
					"address",
					"email",
					"role",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateUserCreateModel": {
				"properties": {
					"fname": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"lname": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"email": {
						"type": "string",
						"pattern": "^\\w+([\\.-]?\\w+)*@\\w+([\\.-]?\\w+)*(\\.\\w{2,3})+$"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"phone": {
						"type": "string",
						"minLength": 9,
						"pattern": "[0-9]{9,12}"
					},
					"address": {
						"type": "string",
						"minLength": 5
					},
					"password": {
						"type": "string",
						"minLength": 6
					}
				},
				"required": [
					"fname",
					"lname",
					"email",
					"phone",
					"address",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateUserUpdateModel": {
				"properties": {
					"fname": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"lname": {
						"type": "string",
						"pattern": "^(?!\\s*$).+"
					},
					"email": {
						"type": "string",
						"pattern": "^\\w+([\\.-]?\\w+)*@\\w+([\\.-]?\\w+)*(\\.\\w{2,3})+$"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"phone": {
						"type": "string",
						"minLength": 9,
						"pattern": "[0-9]{9,12}"
					},
					"address": {
						"type": "string",
						"minLength": 5
					}
				},
				"required": [
					"fname",
					"lname",
					"email",
					"phone",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "cookie",
				"name": "token"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "belle-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "LinhTran"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUsers"
								}
							}
						}
					}
				},
				"description": "Get all users",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UserRole"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UserField"
						}
					},
					{
						"in": "query",
						"name": "change",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Change"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithOrders"
								}
							}
						}
					}
				},
				"description": "Create new user with admin permission",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateUserCreateModel"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithOrders"
								}
							}
						}
					}
				},
				"description": "Get user and user's order list by user id",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor",
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithOrders"
								}
							}
						}
					}
				},
				"description": "Update user info",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"editor",
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ValidateUserUpdateModel"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}